dapr_cosmos_mcp_server/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
httpx[socks,http2]
requests
sseclient-py
azure-cosmos
azure-identity
python-dotenv
aiohttp
dapr-ext-fastapi
isodate
aiohttp
azure-servicebus==7.12.3